pyproject.toml
[tool.poetry]
name = "music-catalogue"
version = "0.1.0"
description = ""
authors = ["Lara <94401584+lara-bellatin@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.124.0"
uvicorn = "^0.38.0"
supabase = "^2.25.0"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
